Pivot controls and example auto, cleanup
This adds a first attempt at setting up controls for the intake and
shooting positions, but will need testing and refinement. There are
several TODOs for tuning the encoder, PID gains, and angles for each
intake and shooting position.

I also added a really basic Auto command as a starting point. I
confirmed in the simulator that this runs in sequence when switching to
Auto, but that's about all.